backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
#passlib[bcrypt]


# Framework y servidor
fastapi==0.103.1
uvicorn==0.23.2

# Validación y tipos
pydantic==2.3.0
typing_extensions>=4.7.1

# Base de datos
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0

# Procesamiento de datos y reportes
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlsxwriter==3.1.2

# Comunicación y archivos
python-multipart==0.0.6
python-dotenv==1.0.0

# Logging y utilidades
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Validación y utilidades adicionales
email-validator==2.0.0
python-dateutil==2.8.2

# Testing (opcional)
pytest==7.4.0
pytest-asyncio==0.21.1
requests==2.31.0

# Utilidades
protobuf==4.24.3
jinja2==3.1.2

